finally fixed Serial Start button. Now properly starts and stops serial streaming. fixed 32bit rollover in sd card writing. working on introducing more control over sd card behavior

+++ boardcode/board_backend.hpp
@@ -25,13 +25,13 @@ class BoardBackend {
     const char*   m_SSID;
     const char*   m_Password;
     char          m_Msg[200];
-    uint64_t      m_BaseTimeMicros{0};
-    uint64_t      m_LocalSyncMicros{0};
+    int64_t      m_BaseTimeMicros{0};
+    int64_t      m_LocalSyncMicros{0};
     volatile bool m_IsTimeSynced{0};
     volatile bool m_WifiOn{true};
     volatile bool m_LoRaOn{false};
-    static uint32_t m_LastSend;
-    static uint32_t m_LastCleanup;
+    static int64_t m_LastSend;
+    static int64_t m_LastCleanup;
     BoardWifi m_wifi;
     SDCard m_sd;
     RPMCollector                  wheel_rc;
